Add a service about saving images.

+++ src/main/java/com/kht/backend/controller/ImageController.java
@@ -13,7 +13,9 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
@@ -25,23 +27,28 @@ public class ImageController {
     private ImageService imageService;
 
     @RequestMapping(value = "/image", method = POST)
-    public Result saveImage(@RequestParam("image") MultipartFile image) {
-        String imageUrl = null;
+    public Result saveImage(@RequestParam("images") MultipartFile images[]) {
+        List<String> imageUrlList = new ArrayList<>();
         try {
-            if (image != null && !image.isEmpty()) {
-                String name = image.getOriginalFilename();
-                String ext = name.substring(name.lastIndexOf("."));
-                String sha1 = DigestUtils.sha1Hex(image.getInputStream());
-                String fileName = sha1 + ext;
-                String path = System.getProperty("java.io.tmpdir") + File.separator + imageService.sha1ToPath(sha1);
-                imageService.mkdirs(path);
-                FileUtils.copyInputStreamToFile(image.getInputStream(), new File(path, fileName));
-                imageUrl = imageService.uploadImage("http://119.23.239.101:4869/upload", path + File.separator + fileName);
-                if (imageUrl != null && !imageUrl.trim().isEmpty()) {
-                    File file = new File(path, fileName);
-                    file.setWritable(true);
-                    System.gc();
-                    file.delete();
+            if (images != null && images.length > 0) {
+                for (MultipartFile image : images) {
+                    if (image != null && !image.isEmpty()) {
+                        String name = image.getOriginalFilename();
+                        String ext = name.substring(name.lastIndexOf("."));
+                        String sha1 = DigestUtils.sha1Hex(image.getInputStream());
+                        String fileName = sha1 + ext;
+                        String path = System.getProperty("java.io.tmpdir") + File.separator + imageService.sha1ToPath(sha1);
+                        imageService.mkdirs(path);
+                        FileUtils.copyInputStreamToFile(image.getInputStream(), new File(path, fileName));
+                        String imageUrl = imageService.uploadImage("http://119.23.239.101:4869/upload", path + File.separator + fileName);
+                        if (imageUrl != null && !imageUrl.trim().isEmpty()) {
+                            imageUrlList.add(imageUrl);
+                            File file = new File(path, fileName);
+                            file.setWritable(true);
+                            System.gc();
+                            file.delete();
+                        }
+                    }
                 }
             }
         }
@@ -49,7 +56,7 @@ public class ImageController {
             throw new ServiceException(ErrorCode.SERVER_EXCEPTION, "图片上传失败");
         }
         Map<String, Object> resultData = new LinkedHashMap<>();
-        resultData.put("imageUrl", imageUrl);
+        resultData.put("imageUrlList", imageUrlList);
         return Result.OK(resultData).build();
     }
 }
